Move sensor address to struct of ds1631.
Move the sensor address to struct of ds1631 as
we can address up to eight devices on the same
bus.
Remove the configurable device address from the
header file as it's user configurable.
Adapt more functions to pass the struct of ds1631
so we can obtain all information from it.